+++ servicetemplates/radon.blueprints.testing/test_simple/ServiceTemplate.tosca
@@ -4,34 +4,6 @@ metadata:
   targetNamespace: "radon.blueprints.testing"
 topology_template:
   node_templates:
-    ReplaceCategoricalValues_0:
-      type: mltosca.nodes.Preprocessing.ReplaceCategoricalValues
-      metadata:
-        x: "1260"
-        y: "102"
-        displayName: "ReplaceCategoricalValues"
-      properties:
-        column: "Sex"
-      requirements:
-        - dependency:
-            node: ShuffleRows_0
-            relationship: con_DependsOn_2
-            capability: feature
-        - host:
-            node: Pipeline_0
-            relationship: con_HostedOnPipeline_2
-            capability: host
-    Accuracy_0:
-      type: mltosca.nodes.Evaluation.Accuracy
-      metadata:
-        x: "1600"
-        y: "823"
-        displayName: "Accuracy"
-      requirements:
-        - host:
-            node: EvaluationPipeline_0
-            relationship: con_HostedOnEvaluation_0
-            capability: host
     RandomForestClassifier_0:
       type: mltosca.nodes.Sklearn.RandomForestClassifier
       metadata:
@@ -52,6 +24,101 @@ topology_template:
         - host:
             node: Conda_0
             relationship: con_HostedOn_3
+            capability: host
+    Precision_0:
+      type: mltosca.nodes.Evaluation.Precision
+      metadata:
+        x: "1574"
+        y: "900"
+        displayName: "Precision"
+      requirements:
+        - host:
+            node: EvaluationPipeline_0
+            relationship: con_HostedOnEvaluation_1
+            capability: host
+    ShuffleRows_0:
+      type: mltosca.nodes.Preprocessing.ShuffleRows
+      metadata:
+        x: "904"
+        y: "96"
+        displayName: "ShuffleRows"
+      requirements:
+        - dependency:
+            node: RemoveColumns_0
+            relationship: con_DependsOn_1
+            capability: feature
+        - host:
+            node: Pipeline_0
+            relationship: con_HostedOnPipeline_0
+            capability: host
+    Recall_0:
+      type: mltosca.nodes.Evaluation.Recall
+      metadata:
+        x: "2233"
+        y: "914"
+        displayName: "Recall"
+      requirements:
+        - host:
+            node: EvaluationPipeline_0
+            relationship: con_HostedOnEvaluation_3
+            capability: host
+    ReplaceCategoricalValues_1:
+      type: mltosca.nodes.Preprocessing.ReplaceCategoricalValues
+      metadata:
+        x: "1586"
+        y: "94"
+        displayName: "ReplaceCategoricalValues"
+      properties:
+        column: "Embarked"
+      requirements:
+        - dependency:
+            node: ReplaceCategoricalValues_0
+            relationship: con_DependsOn_5
+            capability: feature
+        - host:
+            node: Pipeline_0
+            relationship: con_HostedOnPipeline_3
+            capability: host
+    RemoveColumns_0:
+      type: mltosca.nodes.Preprocessing.RemoveColumns
+      metadata:
+        x: "489"
+        y: "89"
+        displayName: "RemoveColumns"
+      properties:
+        columns: "PassengerId,Name, Cabin,Ticket"
+      requirements:
+        - host:
+            node: Pipeline_0
+            relationship: con_HostedOnPipeline_1
+            capability: host
+    ReplaceCategoricalValues_0:
+      type: mltosca.nodes.Preprocessing.ReplaceCategoricalValues
+      metadata:
+        x: "1260"
+        y: "102"
+        displayName: "ReplaceCategoricalValues"
+      properties:
+        column: "Sex"
+      requirements:
+        - dependency:
+            node: ShuffleRows_0
+            relationship: con_DependsOn_2
+            capability: feature
+        - host:
+            node: Pipeline_0
+            relationship: con_HostedOnPipeline_2
+            capability: host
+    Accuracy_0:
+      type: mltosca.nodes.Evaluation.Accuracy
+      metadata:
+        x: "1312"
+        y: "766"
+        displayName: "Accuracy"
+      requirements:
+        - host:
+            node: EvaluationPipeline_0
+            relationship: con_HostedOnEvaluation_0
             capability: host
     RemoveNanRows_0:
       type: mltosca.nodes.Preprocessing.RemoveNanRows
@@ -83,17 +150,6 @@ topology_template:
             node: Conda_0
             relationship: con_HostedOn_0
             capability: host
-    Precision_0:
-      type: mltosca.nodes.Evaluation.Precision
-      metadata:
-        x: "1931"
-        y: "817"
-        displayName: "Precision"
-      requirements:
-        - host:
-            node: EvaluationPipeline_0
-            relationship: con_HostedOnEvaluation_1
-            capability: host
     SplitTrainTest_0:
       type: mltosca.nodes.Splitting.SplitTrainTest
       metadata:
@@ -114,24 +170,9 @@ topology_template:
     Conda_0:
       type: mltosca.nodes.Environment.Conda
       metadata:
-        x: "702"
-        y: "608"
+        x: "424"
+        y: "562"
         displayName: "Conda"
-    ShuffleRows_0:
-      type: mltosca.nodes.Preprocessing.ShuffleRows
-      metadata:
-        x: "904"
-        y: "96"
-        displayName: "ShuffleRows"
-      requirements:
-        - dependency:
-            node: RemoveColumns_0
-            relationship: con_DependsOn_1
-            capability: feature
-        - host:
-            node: Pipeline_0
-            relationship: con_HostedOnPipeline_0
-            capability: host
     ReadCSV_0:
       type: mltosca.nodes.Reading.ReadCSV
       metadata:
@@ -147,11 +188,22 @@ topology_template:
             node: Conda_0
             relationship: con_HostedOn_1
             capability: host
+    F1_0:
+      type: mltosca.nodes.Evaluation.F1
+      metadata:
+        x: "1937"
+        y: "969"
+        displayName: "F1"
+      requirements:
+        - host:
+            node: EvaluationPipeline_0
+            relationship: con_HostedOnEvaluation_2
+            capability: host
     EvaluationPipeline_0:
       type: mltosca.nodes.Evaluation.EvaluationPipeline
       metadata:
-        x: "1785"
-        y: "666"
+        x: "1814"
+        y: "662"
         displayName: "EvaluationPipeline"
       requirements:
         - dependency:
@@ -161,36 +213,6 @@ topology_template:
         - host:
             node: Conda_0
             relationship: con_HostedOn_4
-            capability: host
-    ReplaceCategoricalValues_1:
-      type: mltosca.nodes.Preprocessing.ReplaceCategoricalValues
-      metadata:
-        x: "1586"
-        y: "94"
-        displayName: "ReplaceCategoricalValues"
-      properties:
-        column: "Embarked"
-      requirements:
-        - dependency:
-            node: ReplaceCategoricalValues_0
-            relationship: con_DependsOn_5
-            capability: feature
-        - host:
-            node: Pipeline_0
-            relationship: con_HostedOnPipeline_3
-            capability: host
-    RemoveColumns_0:
-      type: mltosca.nodes.Preprocessing.RemoveColumns
-      metadata:
-        x: "489"
-        y: "89"
-        displayName: "RemoveColumns"
-      properties:
-        columns: "PassengerId,Name, Cabin,Ticket"
-      requirements:
-        - host:
-            node: Pipeline_0
-            relationship: con_HostedOnPipeline_1
             capability: host
   relationship_templates:
     con_HostedOnPipeline_4:
@@ -232,4 +254,8 @@ topology_template:
     con_HostedOnEvaluation_1:
       type: mltosca.relationships.HostedOnEvaluation
     con_HostedOnEvaluation_0:
+      type: mltosca.relationships.HostedOnEvaluation
+    con_HostedOnEvaluation_3:
+      type: mltosca.relationships.HostedOnEvaluation
+    con_HostedOnEvaluation_2:
       type: mltosca.relationships.HostedOnEvaluation
